support multi-AZ cluster
We need to support a multi-az rollout for HA.
Due to our cidr block allocation strategy, we can't use the
Apollo-upstream scheme of creating cidr blocks based on a counter.
Instead, we will pre-compute blocks and save them in terraform.tfvars.

Note that elasticsearch isn't setup in HA mode yet.