Use public cloud for digital ocean and secure with TLS
"Private network" on digital ocean is really a shared private network.

It means all other hosts in datacenter, even of other users, can access
deployed nodes on digitalocean. It means deploying on public ips is
not less secure than deploying on private ones.

In this setup we additionaly secure etcd with wildcard tls certs.

As an additional bonus, we can not deploy servers across datacenters.

Also, we generate initial ssh key for digital ocean dynamically, for
better security, and export it to "id_rsa" file for easy ssh access.